backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
openai==1.56.0
pydantic==2.9.2

tree-sitter==0.21.3
tree-sitter-languages==1.10.2
PyJWT==2.8.0
cryptography==42.0.8
